(AccountAvatar) move folder to component-library dir
