+++ app/component-library/components/AccountAvatar/AccountAvatar.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 
-import { AvatarSize } from '../../Base/BaseAvatar';
+import { BaseAvatarSize } from '../BaseAvatar';
 import AccountAvatar, { AccountAvatarType } from '.';
 
 const accountAddress = 'sitean';
@@ -10,7 +10,7 @@ describe('BaseAvatar', () => {
   it('should render correctly', () => {
     const wrapper = shallow(
       <AccountAvatar
-        size={AvatarSize.ExtraLarge}
+        size={BaseAvatarSize.Xl}
         type={AccountAvatarType.JazzIcon}
         accountAddress={accountAddress}
       />,
