+++ app/component-library/components/AccountAvatar/README.md
