+++ app/component-library/components/AccountAvatar/index.ts
